+++ src/index.js
@@ -9,16 +9,27 @@ import { UserContextProvider } from "./context/userContext";
 
 import Header from "./components/header";
 import Footer from "./components/footer";
+import styled from "styled-components";
+
+const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+`;
+
 
 const root = ReactDOM.createRoot(document.querySelector("#root"));
 root.render(
   <UserContextProvider>
     <CartContextProvider>
       <BrowserRouter>
-        <Header />
-        <App />
-        <Footer />
+        <Container>
+          <Header />
+          <App />
+          <Footer />
+        </Container>
       </BrowserRouter>
     </CartContextProvider>
   </UserContextProvider>
 );
+
