+++ src/views/Cart.jsx
@@ -6,11 +6,11 @@ import { useUserContext } from "../context/userContext";
 
 import styled from "styled-components";
 
-
 import theme from "../lib/themes";
 
-import loginEmail from "../functions/loginEmail";
 import createCheckoutSession from "../functions/createCheckoutSession";
+
+import Modal from "../components/modal";
 
 const Cart = () => {
   const { cart } = useCartContext();
@@ -21,31 +21,10 @@ const Cart = () => {
     user ? createCheckoutSession(user.uid, cart) : setIsModal(true);
   }
 
-  function login(e) {
-    e.preventDefault();
-    const email = e.target.email.value;
-    const password = e.target.password.value;
-    loginEmail(email, password);
-    setIsModal(false);
-    createCheckoutSession(user.uid, cart);
-  }
   return (
     <Container>
       {isModal ? (
-        <BlurBackground>
-          <Card>
-            <h3>Sign in or Register</h3>
-            <ModalForm onSubmit={(e) => login(e)}>
-              <Input type="text" name="email" placeholder="E-mail" />
-              <Input type="password" name="password" placeholder="Password" />
-              <Button
-                title="Login"
-                backgroundColor={theme.color.primary}
-                colorLabel={theme.color.white}
-              />
-            </ModalForm>
-          </Card>
-        </BlurBackground>
+        <Modal setIsModal={setIsModal} cart={cart} user={user} />
       ) : (
         <div style={{ backgroundColor: "blue" }}></div>
       )}
@@ -55,10 +34,10 @@ const Cart = () => {
           to="/"
           style={{
             color: `${theme.color.white}`,
+            cursor: "pointer",
             padding: 14,
             fontSize: `${theme.fontSize.small}`,
             borderRadius: 6,
-            cursor: "pointer",
             textDecoration: "none",
             backgroundColor: `${theme.color.primary}`,
           }}
@@ -87,51 +66,11 @@ const Container = styled.div`
   position: relative;
 `;
 
-const BlurBackground = styled.div`
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100vh;
-  z-index: 30;
-  display: flex;
-  position: absolute;
-  min-width: 288px;
-  align-items: center;
-  justify-content: center;
-  backdrop-filter: blur(8px);
-  background-color: rgba(0, 0, 0, 0.5);
-`;
-
-const Card = styled.section`
-  width: 90%;
-  height: 300px;
-  margin: 16px auto;
-  display: flex;
-  min-width: 288px;
-  max-width: 600px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.16);
-  align-items: center;
-  border-radius: 6px;
-  flex-direction: column;
-  justify-content: center;
-  background-color: ${theme.color.white};
-`;
-
-const ModalForm = styled.form`
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-`;
-
-const Input = styled.input`
-  margin: 10px 0;
-`;
-
 const CartContainer = styled.section`
   width: 90%;
   margin: 16px auto;
-  display: flex;
   cursor: pointer;
+  display: flex;
   min-width: 288px;
   max-width: 400px;
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.16);
